+++ src/game/player/lib/flush.tsx
@@ -1,14 +1,18 @@
-import React, {useEffect, useReducer} from "react";
+import React, {useEffect, useState} from "react";
 
 export function useFlush(deps: React.DependencyList = []): [
     () => void,
     number
 ] {
-    const [count, update] = useReducer(x => x + 1, 0);
+    const [count, setCount] = useState<number>(0);
 
     useEffect(() => {
         update();
     }, deps);
+
+    function update() {
+        setCount((count) => count + 1);
+    }
 
     return [update, count];
 }