+++ src/util/data.ts
@@ -1040,3 +1040,13 @@ export function ThenableAll(thenables: Thenable<any>[]): Promise<any[]> {
         return thenable;
     }));
 }
+
+export function onlyValidFields<T extends Record<string, any>>(obj: T): Partial<T> {
+    const result: Partial<T> = {};
+    for (const key in obj) {
+        if (obj[key] !== undefined) {
+            result[key] = obj[key];
+        }
+    }
+    return result;
+}